refactor: consolidate enrollment and promotion logic, improve validation
Major refactoring to improve code organization, validation, and maintainability
across the program enrollment tools.

Code Organization:
- Extract shared promotion and enrollment functions from doctype files to utils.py

Validation Improvements:
- Add before_save() validation to both enrollment tools:
  - Prevent same academic year/term for current and new fields
  - Validate promotion rules are defined
  - Prevent same class/stream in promotion rules
- Add mandatory_depends_on condition for promotion_rules_engine field
- Add depends_on condition for section_break_5 in enhanced tool

UI/UX Enhancements:
- Refactor set_stream_filters() in automated tool JS:
  - Convert from standalone function to form method
  - Trigger automatically on academic_year and program changes
  - Clear promotion_rules_engine table when program changes
- Hide enroll_students button in automated tool
- Add freeze message during enrollment process in enhanced tool

Functional Improvements:
- Pass enrollment_date parameter to enrollment function instead of
  hardcoding current date
- Improve update_enrollment_tool() with better academic year filtering
  using date range queries
- Fix create_academic_year() to use dynamic year calculation instead
  of hardcoded "2029"
- Use proper date utility functions (get_year_start, get_year_ending, etc.)

Bug Fixes:
- Improve error handling in enrollment function with try/except per student

This refactoring reduces code duplication, improves maintainability, and
adds important validation to prevent data integrity issues.
